# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Task Manager API",
    "version": "1.0.0",
    "description": "Comprehensive API documentation for the AI Task Manager SaaS platform with task management, AI features, team management, gamification, analytics, and more."
  },
  "servers": [
    { "url": "http://localhost:3000/api" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "File": {
        "type": "object",
        "properties": {
          "filename": { "type": "string", "description": "Unique filename on server" },
          "originalName": { "type": "string", "description": "Original filename from user" },
          "url": { "type": "string", "description": "File access URL" },
          "fileType": { "type": "string", "description": "File extension" },
          "fileSize": { "type": "number", "description": "File size in bytes" },
          "mimeType": { "type": "string", "description": "MIME type of the file" },
          "uploadedBy": { "type": "string", "description": "User ID who uploaded the file" },
          "task": { "type": "string", "description": "Associated task ID" },
          "project": { "type": "string", "description": "Associated project ID" },
          "team": { "type": "string", "description": "Associated team ID" },
          "department": { "type": "string", "description": "Associated department ID" },
          "conversation": { "type": "string", "description": "Associated conversation ID" },
          "category": { "type": "string", "enum": ["document", "image", "video", "audio", "archive", "other"], "description": "File category" },
          "tags": { "type": "array", "items": { "type": "string" }, "description": "File tags" },
          "description": { "type": "string", "description": "File description" },
          "isPublic": { "type": "boolean", "description": "Whether file is publicly accessible" },
          "isShared": { "type": "boolean", "description": "Whether file is shared with others" },
          "sharedWith": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "user": { "type": "string", "description": "User ID shared with" },
                "permission": { "type": "string", "enum": ["view", "edit", "admin"], "description": "Permission level" },
                "sharedAt": { "type": "string", "format": "date-time", "description": "When file was shared" }
              }
            },
            "description": "List of users file is shared with"
          },
          "downloadCount": { "type": "number", "description": "Number of times file was downloaded" },
          "lastDownloaded": { "type": "string", "format": "date-time", "description": "Last download timestamp" },
          "isDeleted": { "type": "boolean", "description": "Whether file is soft deleted" },
          "deletedAt": { "type": "string", "format": "date-time", "description": "Deletion timestamp" },
          "deletedBy": { "type": "string", "description": "User ID who deleted the file" },
          "createdAt": { "type": "string", "format": "date-time", "description": "Upload timestamp" },
          "updatedAt": { "type": "string", "format": "date-time", "description": "Last update timestamp" }
        },
        "required": ["filename", "originalName", "url", "uploadedBy"]
      },
...
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
...
  "paths": {
...
    "/chat/conversations": {
      "get": {
        "tags": ["Chat"],
        "summary": "Get all conversations for the authenticated user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of conversations retrieved successfully" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      },
      "post": {
        "tags": ["Chat"],
        "summary": "Create new conversation",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "description": "Conversation name" },
                  "type": { "type": "string", "enum": ["direct", "group", "team", "project"], "description": "Conversation type" },
                  "participants": { "type": "array", "items": { "type": "string" }, "description": "Array of user IDs" },
                  "projectId": { "type": "string", "description": "Project ID for project conversations" },
                  "teamId": { "type": "string", "description": "Team ID for team conversations" },
                  "taskId": { "type": "string", "description": "Task ID for task conversations" }
                },
                "required": ["type", "participants"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Conversation created successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
    "/chat/conversations/{id}": {
      "delete": {
        "tags": ["Chat"],
        "summary": "Delete conversation",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Conversation deleted successfully" },
          "403": { "description": "Access denied" },
          "404": { "description": "Conversation not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/chat/messages/{convId}": {
      "get": {
        "tags": ["Chat"],
        "summary": "Get messages by conversation ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "convId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": {
          "200": { "description": "Messages retrieved successfully" },
          "403": { "description": "Access denied" },
          "404": { "description": "Conversation not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
    "/chat/messages/{id}": {
      "put": {
        "tags": ["Chat"],
        "summary": "Edit message",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string", "description": "New message content" }
                },
                "required": ["message"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Message updated successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Access denied" },
          "404": { "description": "Message not found" },
          "500": { "description": "Internal server error" }
        }
      },
      "delete": {
        "tags": ["Chat"],
        "summary": "Delete message",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Message deleted successfully" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Access denied" },
          "404": { "description": "Message not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
    "/chat/upload": {
      "post": {
        "tags": ["Chat"],
        "summary": "Upload file to chat",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary", "description": "File to upload" },
                  "conversationId": { "type": "string", "description": "Conversation ID" },
                  "description": { "type": "string", "description": "File description" }
                },
                "required": ["file", "conversationId"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "File uploaded successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Access denied or file uploads not allowed" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/chat/files/{convId}": {
      "get": {
        "tags": ["Chat"],
        "summary": "Get chat files for a conversation",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "convId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Chat files retrieved successfully" },
          "403": { "description": "Access denied" },
          "404": { "description": "Conversation not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/chat/files/{id}": {
      "delete": {
        "tags": ["Chat"],
        "summary": "Delete chat file",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Chat file deleted successfully" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Access denied" },
          "404": { "description": "File not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/chat/{taskId}": {
      "get": {
        "tags": ["Chat"],
        "summary": "Get all chat messages for a task",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of chat messages." },
          "401": { "description": "Unauthorized." }
        }
      },
      "post": {
        "tags": ["Chat"],
        "summary": "Send a chat message within a task",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string" }
                },
                "required": ["message"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Message sent." },
          "400": { "description": "Validation error." },
          "401": { "description": "Unauthorized." }
        }
      }
    },
...
    "/files/{id}": {
      "get": {
        "tags": ["File Management"],
        "summary": "Get file by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "File ID" }
        ],
        "responses": {
          "200": {
            "description": "File retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "file": { "$ref": "#/components/schemas/File" }
                  }
                }
              }
            }
          },
          "403": { "description": "Access denied" },
          "404": { "description": "File not found" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      },
      "delete": {
        "tags": ["File Management"],
        "summary": "Delete file",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "File ID" }
        ],
        "responses": {
          "200": {
            "description": "File deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "fileId": { "type": "string" }
                  }
                }
              }
            }
          },
          "403": { "description": "Access denied" },
          "404": { "description": "File not found" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
  }
}